+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Pass selected skuId in query string.
+
 ## [0.2.1] - 2021-01-13
 
 ### Fixed