+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added 
+- `crowdin.yml` for crowdin integration.
+
 ## [0.2.3] - 2021-01-29
 ### Fixed
 - Remove unavailable sellers from the list of sellers.