+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [Unreleased]
 
 ### Added 
-- `crowdin.yml` for crowdin integration.
+- Crowdin.yml for crowdin integration.
 
 ## [0.2.3] - 2021-01-29
 ### Fixed